Removed the V0 code from the API
