--- api/app/signals/apps/api/v0/fields.py
@@ -1,74 +1,0 @@
-"""
-Signals API V0 (legacy) custom serializer fields.
-"""
-
-from collections import OrderedDict
-
-from rest_framework import serializers
-
-
-class SignalLinksField(serializers.HyperlinkedIdentityField):
-    """
-    Return authorized url. handy for development.
-    """
-
-    def to_representation(self, value):
-        request = self.context.get('request')
-
-        result = OrderedDict([
-            ('self', dict(
-                href=self.get_url(value, "signal-auth-detail", request, None))
-             ),
-        ])
-
-        return result
-
-
-class SignalUnauthenticatedLinksField(serializers.HyperlinkedIdentityField):
-    """
-    Return url based on UUID instead of normal database id
-    """
-    lookup_field = 'signal_id'
-
-
-class StatusLinksField(serializers.HyperlinkedIdentityField):
-    """
-    Return authorized url. handy for development.
-    """
-
-    def to_representation(self, value):
-        request = self.context.get('request')
-
-        result = OrderedDict([
-            ('self', dict(
-                href=self.get_url(value, "status-auth-detail", request, None))
-             ),
-        ])
-
-        return result
-
-
-class CategoryLinksField(serializers.HyperlinkedIdentityField):
-
-    def to_representation(self, value):
-        request = self.context.get('request')
-
-        result = OrderedDict([
-            ('self', dict(
-                href=self.get_url(value, "category-auth-detail", request, None))
-             ),
-        ])
-
-        return result
-
-
-class PriorityLinksField(serializers.HyperlinkedIdentityField):
-
-    def to_representation(self, value):
-        request = self.context.get('request')
-
-        result = OrderedDict([
-            ('self', dict(href=self.get_url(value, 'priority-auth-detail', request, None))),
-        ])
-
-        return result
