+++ api/app/signals/urls.py
@@ -5,8 +5,6 @@ from django.views.decorators.csrf import csrf_exempt
 
 from signals.apps.graphql.views import SIAGraphQLView
 
-# from signals.apps.users.v0.views import UserMeView
-
 urlpatterns = [
     path('status/', include('signals.apps.health.urls')),
 
@@ -14,16 +12,8 @@ urlpatterns = [
     path('signals/', include('signals.apps.api.urls')),
     path('signals/admin/', admin.site.urls),
 
-    # Disbaled V0
-    # DEPRECATED url route for `auth/me`. Should be fixed in the frontend before we can remove
-    # this endpoint here.
-    # path('signals/auth/me/', UserMeView.as_view()),
-    # path('signals/user/auth/me/', UserMeView.as_view()),
-
     # SOAP stand-in endpoints
     path('signals/sigmax/', include('signals.apps.sigmax.urls')),
-
-    # path('signals/experimental/reporting/', include('signals.apps.reporting.urls'))
 
     # GraphQL endpoint
     path('signals/graphql', csrf_exempt(SIAGraphQLView.as_view(graphiql=True))),
