--- api/app/tests/apps/api/v0/request_data/category_auth_post.json
@@ -1,5 +1,0 @@
-{
-    "main": "Overlast op het water",
-    "sub": "Overlast op het water - snel varen",
-    "priority": 0
-}
