--- api/app/tests/apps/api/v0/request_data/location_auth_post.json
@@ -1,14 +1,0 @@
-{
-	"geometrie": {
-	    "type": "Point",
-	    "coordinates": [
-		4.898466,
-		52.361585
-	    ]
-	},
-	"_signal": "/signals/signal/8/",
-	"stadsdeel": "A",
-	"buurt_code": "aaa1",
-	"address": {},
-	"extra_properties": {}
-}
