--- api/app/tests/apps/api/v0/request_data/status_auth_post.json
@@ -1,10 +1,0 @@
-{
-	"text": "24 123 123 123 jaa",
-	"user": "stephan@apibouwer.nl",
-	"target_api": "sigma",
-	"state": "i",
-	"extern": true,
-	"extra_properties": {},
-	"_signal": "/signals/signal/8/"
-}
-
