+++ api/app/tests/apps/api/v1/test_ml_predict_endpoint.py
@@ -1,6 +1,7 @@
-from unittest import skip
-from unittest.mock import patch
+from unittest import mock, skip
+from unittest.mock import MagicMock, patch
 
+from django.core.exceptions import ValidationError
 from django.http import Http404
 from rest_framework.exceptions import APIException
 
@@ -10,6 +11,7 @@ from tests.apps.signals.factories import CategoryFactory
 from tests.test import SignalsBaseApiTestCase
 
 
+@skip('V1 Disabled')
 class TestMlPredictCategory(SignalsBaseApiTestCase):
     test_host = 'http://testserver'
     endpoint = '/signals/v1/public/category/prediction'
@@ -51,7 +53,6 @@ class TestMlPredictCategory(SignalsBaseApiTestCase):
             created_by='someone@example.com',
         )
 
-    @skip('V1 Disabled')
     @patch('signals.apps.api.v1.public.views.MLPredictCategoryView._ml_predict')
     def test_predict(self, patched):
         patched.return_value = self.link_test_subcategory
@@ -65,7 +66,6 @@ class TestMlPredictCategory(SignalsBaseApiTestCase):
 
         self.assertEqual(response_json['_links']['self']['href'], self.link_test_subcategory)
 
-    @skip('V1 Disabled')
     @patch('signals.apps.api.v1.public.views.MLPredictCategoryView._ml_predict')
     def test_predict_translated(self, patched):
         patched.return_value = self.link_test_subcategory_translated
@@ -80,7 +80,6 @@ class TestMlPredictCategory(SignalsBaseApiTestCase):
         # This should be the translated category URL
         self.assertEqual(response_json['_links']['self']['href'], self.link_test_subcategory)
 
-    @skip('V1 Disabled')
     @patch('signals.apps.api.v1.public.views.MLPredictCategoryView._ml_predict')
     def test_predict_overig(self, patched):
         patched.return_value = self.link_test_subcategory_overig
@@ -94,7 +93,6 @@ class TestMlPredictCategory(SignalsBaseApiTestCase):
 
         self.assertEqual(response_json['_links']['self']['href'], self.link_test_subcategory_overig)
 
-    @skip('V1 Disabled')
     @patch('signals.apps.api.v1.public.views.MLPredictCategoryView._ml_predict')
     def test_predict_overig_none(self, patched):
         patched.return_value = None
@@ -108,7 +106,6 @@ class TestMlPredictCategory(SignalsBaseApiTestCase):
 
         self.assertEqual(response_json['_links']['self']['href'], self.link_test_subcategory_overig)
 
-    @skip('V1 Disabled')
     @patch('signals.apps.api.v1.public.views.MLPredictCategoryView._ml_predict',
            side_effect=Http404)
     def test_predict_404(self, patched):
@@ -117,7 +114,6 @@ class TestMlPredictCategory(SignalsBaseApiTestCase):
 
         self.assertEqual(response.status_code, 404)
 
-    @skip('V1 Disabled')
     @patch('signals.apps.api.v1.public.views.MLPredictCategoryView._ml_predict',
            side_effect=APIException)
     def test_predict_httpresponseservererror(self, patched):
@@ -126,9 +122,82 @@ class TestMlPredictCategory(SignalsBaseApiTestCase):
 
         self.assertEqual(response.status_code, 500)
 
-    @skip('V1 Disabled')
     def test_predict_invalid_request(self):
         data = {}
         response = self.client.get(self.endpoint, data=data, format='json')
 
         self.assertEqual(response.status_code, 400)
+
+
+class TestLegacyMlPredictCategory(SignalsBaseApiTestCase):
+    test_host = 'http://testserver'
+    endpoint = '/signals/category/prediction'
+
+    def setUp(self):
+        self.test_parentcategory_overig = Category.objects.get(slug='overig', parent__isnull=True)
+        self.test_subcategory_overig = Category.objects.get(parent=self.test_parentcategory_overig, slug='overig',
+                                                            parent__isnull=False)
+        self.link_test_subcategory_overig = f'{self.test_host}/signals/v1/public/terms/categories/' \
+                                            f'{self.test_subcategory_overig.parent.slug}/sub_categories/' \
+                                            f'{self.test_subcategory_overig.slug,}'
+
+        self.test_subcategory = CategoryFactory.create()
+        self.link_test_subcategory = f'{self.test_host}/signals/v1/public/terms/categories/' \
+                                     f'{self.test_subcategory.parent.slug}/sub_categories/{self.test_subcategory.slug}'
+
+        self.test_subcategory_translated = CategoryFactory.create(is_active=False)
+        self.link_test_subcategory_translated = f'{self.test_host}/signals/v1/public/terms/categories/' \
+                                                f'{self.test_subcategory_translated.parent.slug}/sub_categories/' \
+                                                f'{self.test_subcategory_translated.slug}'
+
+        self.link_test_subcategory_translation = CategoryTranslation.objects.create(
+            old_category=self.test_subcategory_translated,
+            new_category=self.test_subcategory,
+            text='For testing purposes we translate this category',
+            created_by='someone@example.com',
+        )
+
+    @patch('signals.apps.api.v1.views.LegacyMlPredictCategoryView.ml_tool_client.predict')
+    def test_predict(self, patched_ml_tool_client):
+        response = mock.Mock()
+        response.status_code = 200
+        response.json = MagicMock(return_value={
+            'hoofdrubriek': [[self.link_test_subcategory], [0.5]],
+            'subrubriek': [[self.link_test_subcategory], [0.5]]
+        })
+        patched_ml_tool_client.return_value = response
+
+        data = {'text': 'Give me the subcategory'}
+        response = self.client.post(self.endpoint, data=data, format='json')
+
+        self.assertEqual(response.status_code, 200)
+
+        response_json = response.json()
+        self.assertEqual(response_json['subrubriek'][0][0], self.link_test_subcategory)
+
+    @patch('signals.apps.api.v1.views.LegacyMlPredictCategoryView.ml_tool_client.predict')
+    def test_predict_translated(self, patched_ml_tool_client):
+        response = mock.Mock()
+        response.status_code = 200
+        response.json = MagicMock(return_value={
+            'hoofdrubriek': [[self.link_test_subcategory_translated], [0.5]],
+            'subrubriek': [[self.link_test_subcategory_translated], [0.5]]
+        })
+        patched_ml_tool_client.return_value = response
+
+        data = {'text': 'Give me the subcategory, because of translations'}
+        response = self.client.post(self.endpoint, data=data, format='json')
+
+        self.assertEqual(response.status_code, 200)
+
+        response_json = response.json()
+        self.assertEqual(response_json['subrubriek'][0][0], self.link_test_subcategory)
+
+    @patch('signals.apps.api.v1.views.LegacyMlPredictCategoryView.ml_tool_client.predict',
+           side_effect=ValidationError('error'))
+    def test_predict_validation_error(self, *args):
+        data = {'text': 'Give me the subcategory'}
+        response = self.client.post(self.endpoint, data=data, format='json')
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(response.json()[0], 'error')
